# backend/api_docs.yaml
openapi: 3.0.2
info:
  title: Backend API v1
  version: 1.0.0
servers:
  - url: http://localhost/api/v1
...
paths:
...
  /auth/refresh:
    get:
      tags:
        - auth
      summary: Used to create an access token.
      description: |-
        Shouldn't be used directly  
        Trying to access a restricted data without (or with an invalidated) will automatically redirect to this
      operationId: RefreshGet
      parameters:
        - name: from
          in: query
          required: false
          description: the url to redirect back after a Successful refresh
          schema:
            type: string
            example: https%3A%2F%2Fexample.com%2Fexample
      responses:
        200:
          description: Successfully fetched an access token
        302:
          description: Redirecting back to the path specified in the "from" parameter
        401:
          description: User not logged in
        429:
          description: You've been rate limited
        500:
          description: Internal error
  /auth/log_in:
    post:
      tags:
        - auth
      summary: Used to log in
      description: |-
        Used for logging in.  
        Sets all the necessary auth cookies
      operationId: LogInPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  $ref: "#/components/schemas/Username"
                password:
                  $ref: "#/components/schemas/Password"
      responses:
        200:
          description: Successfully logged in
        400:
          description: Invalid schema
        401:
          description: Invalid username or password
        429:
          description: You've been rate limited
        500:
          description: Internal error
  /auth/log_out:
    get:
      tags:
        - auth
      summary: used to log out
      description: Used for logging out.
      operationId: LogOutGet
      responses:
        200:
          description: Successfully logged out
        401:
          description: User wasn't logged in, in the first place
        429:
          description: You've been rate limited
        500:
          description: Internal error
...
components:
  schemas:
    Username:
      type: string
      description: |-
        - 5 - 25 characters
        - at least one letter
        - no space allowed
      example: BestChef
    Password:
      type: string
      description: |-
        - 7 - 50 characters
        - 1 upper case letter
        - 1 lower case letter
        - 1 special from this list: !@#$%^&*()-_=+[]{}\\|<>,./?
      example: Dupa123!
...